set up PBF
trying out yet another fluid simulation technique based on Mueller "Position Based Fluid" -- my SPH was dogshit and the FLIP looked great but was a pain to parallelize. PBF looks really good so far -- behavior a lot better, like the FLIP, and decently fast when done through a compute shader. still need to add viscosity and change some of the smoothing kernels. (also not sure about the correction term he suggests to prevent clumping, but we'll wait until everything else is in place).